refactor: centralize formatAddress
