+++ src/components/TransactionDetails.tsx
@@ -3,7 +3,8 @@ import { TransactionReceipt, TransactionResponse } from "ethers";
 
 import { Button } from "./ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
-import { getTransaction, getTransactionReceipt, formatAddress, formatTimestamp } from "../lib/blockchain";
+import { getTransaction, getTransactionReceipt, formatTimestamp } from "../lib/blockchain";
+import { formatAddress } from "../lib/format";
 import { TransactionData } from "./transaction/TransactionData";
 import { TransactionLogs } from "./transaction/TransactionLogs";
